requirement_airflow.txt
yfinance
happybase
apache-airflow
apache-airflow-providers-apache-spark
pyspark
